+++ src/components/quadrant-stats.tsx
@@ -1,5 +1,6 @@
-import { Card, CardHeader, CardTitle, CardContent } from '@/components/ui/card';
-import type { QuadrantData } from '@/types';
+// filepath: /Users/briandouglas/code/contributor.info/src/components/quadrant-stats.tsx
+import { Card, CardHeader, CardTitle, CardContent } from "@/components/ui/card";
+import type { QuadrantData } from "@/types";
 
 interface QuadrantStatsProps {
   data: QuadrantData[];
@@ -14,7 +15,9 @@ export function QuadrantStats({ data }: QuadrantStatsProps) {
             <CardTitle className="text-lg">{quadrant.name}</CardTitle>
           </CardHeader>
           <CardContent>
-            <div className="text-2xl font-bold">{quadrant.percentage.toFixed(1)}%</div>
+            <div className="text-2xl font-bold">
+              {quadrant.percentage.toFixed(1)}%
+            </div>
             <p className="text-sm text-muted-foreground">
               {quadrant.count} PRs
             </p>
